quitado el valor unico a nombre de tienda, y la pagina principal de bo-portal esta decente

+++ application/views/portal-bo/index.php
@@ -7,7 +7,7 @@
   <h1><?php echo $titulo; ?></h1>
 
   <h3>Categorías de la web</h3>
-  <a href='admin/categorias' class="btn btn-default" id="btn-categorias">Administrar categorías de los productos</a>
+  <a href='admin/categorias' class="btn btn-default" id="btn-categorias"><i class="glyphicon glyphicon-cog"></i>Administrar categorías de los productos</a>
 
   <h3>Lista de tiendas</h3>
 
@@ -19,7 +19,7 @@
 		<thead style="background-color:grey"> 
 			<tr> 
 				<th>ID</th> 
-				<th>Nombre <font color="red">(debe ser único)</font></th>
+				<th>Nombre</th>
 				<th>Localización</th>
 				<th>Fecha apertura</th> 
 				<th>Info contacto</th>
@@ -33,6 +33,7 @@
 				<tr> 
 					<th scope="row">
 						<span class="idtienda"><?php echo $tienda->id; ?></span>
+						<span class="error" style="color:red;display:none">Todos los campos son obligatorios</span>
 					</th> 
 					<td>
 						<label><?php echo $tienda->nombre; ?></label>
@@ -44,7 +45,7 @@
 					</td>
 					<td>
 						<label><?php echo $tienda->fechaapertura; ?></label>
-						<input placeholder="Fecha de apertura..." class="form-control" type="text" style="display:none" />
+						<input placeholder="Fecha de apertura..." class="form-control" type="date" style="display:none" />
 					</td> 
 					<td>
 						<label><?php echo $tienda->infocontacto; ?></label>
@@ -64,7 +65,7 @@
 			?>
 			<tr> 
 					<th scope="row">
-						
+						<span class="error" style="color:red;display:none"></span>
 					</th> 
 					<td>
 						<input placeholder="Nombre..." class="form-control" type="text" />
@@ -73,7 +74,7 @@
 						<input placeholder="Localización..." class="form-control" type="text" />
 					</td>
 					<td>
-						<input placeholder="Fecha de apertura..." class="form-control" type="text"  />
+						<input placeholder="Fecha de apertura..." class="form-control" type="date"  />
 					</td> 
 					<td>
 						<input placeholder="Info de contacto..." class="form-control" type="text" />
@@ -96,7 +97,7 @@
 		<thead style="background-color:grey"> 
 			<tr> 
 				<th>ID</th> 
-				<th>Nombre <font color="red">(debe ser único)</font></th>
+				<th>Nombre</th>
 				<th>Localización</th>
 				<th>Fecha apertura</th> 
 				<th>Info contacto</th>
@@ -105,16 +106,16 @@
 		</thead>
 		<tr> 
 					<th scope="row">
-						
+						<span class="error" style="color:red;display:none"></span>
 					</th> 
 					<td>
 						<input placeholder="Nombre..." class="form-control" type="text" />
-					</td>  
+					</td>
 					<td>
 						<input placeholder="Localización..." class="form-control" type="text" />
 					</td>
 					<td>
-						<input placeholder="Fecha de apertura..." class="form-control" type="text"  />
+						<input placeholder="Fecha de apertura..." class="form-control" type="date"  />
 					</td> 
 					<td>
 						<input placeholder="Info de contacto..." class="form-control" type="text" />
@@ -136,7 +137,7 @@ $(document).ready(function() {
     	var estado = $(this).attr("data-estado");
     	var dad = $(this).parent().parent().parent();
     	var labels = dad.find('label');
-    	var inputs = dad.find('input[type="text"]');
+    	var inputs = dad.find('input');
     	var btn = dad.find(".btn-guardar");
 
     	if(estado=="mostrando") //pasamos a la vista de editar
@@ -168,15 +169,17 @@ $(document).ready(function() {
 
 	$('.btn-guardar').click(function () {
 		var dad = $(this).parent().parent().parent();
-		var inputs = dad.find('input[type="text"]');
+		var inputs = dad.find('input');
 		var id = dad.find('.idtienda')[0].innerHTML;
-
+		var errormensaje = dad.find(".error")[0];
+		$(".error").hide();
 		var datos = {}
 		datos.id = id;
 		for(var i=0;i<inputs.length;i++) 
 		{
 	    	var texto = inputs[i].value;
 	    	if(!texto) {
+	    		errormensaje.style.display="inline";
 	    		return; //si hay un campo sin rellenar, no es valido
 	    	}
 	    	if(i==0)
@@ -195,28 +198,44 @@ $(document).ready(function() {
 	        data : datos,
 	        success:function (data) {
 	        	location.reload();
-	        }
+	        },
+	        error: function(data){
+       			alert("fail");
+    		}
 		});
 	});
 
 	 $('#btn-crear').click(function () {
 	 	var dad = $(this).parent().parent();
-	 	var inputs = dad.find('input[type="text"]');
+	 	var inputs = dad.find(':input[type="text"]');
+	 	var errormensaje = dad.find(".error")[0];
+	 	$(".error").hide();
 	 	//comprobamos que los campos esten rellenos
 	 	var datos = {}
         for(var i=0;i<inputs.length;i++) {
         	var input = inputs[i].value;
 	    	if(!input)
+	    	{
+	    		errormensaje.innerHTML="Todos los campos son obligatorios";
+	    		errormensaje.style.display="inline";
 	    		return;
+	    	}
 	    	if(i==0)
 	    		datos.nombre = input;
 	    	if(i==1)
 	    		datos.localizacion = input;
 	    	if(i==2)
-	    		datos.fechaapertura = input;
-	    	if(i==3)
 	    		datos.infocontacto = input;
         }
+        var fecha = dad.find(':input[type="date"]')[0].value;
+        if(!fecha)
+        {
+        	errormensaje.innerHTML="Todos los campos son obligatorios";
+	    	errormensaje.style.display="inline";
+        	return;
+        }
+        else
+        	datos.fechaapertura = fecha;
         $.ajax({
 	        url : 'admin/tiendas/crear',
 	        type : 'post',
@@ -229,14 +248,20 @@ $(document).ready(function() {
 	 });
 
 	 $('.btn-borrar').click(function () {
+
 	 	var id = $(this).parent().parent().parent().find('.idtienda')[0].innerHTML;
-	 	$.ajax({
-	        url : 'admin/tiendas/' + id + '/borrar',
-	        type : 'delete',
-	        success:function (data) {
-	        	location.reload();
-	        }
-		});
+	 	bootbox.confirm("¿Estás seguro? Perderás todos los productos que sean de esta tienda.", function(seguro) {
+			if(seguro)
+			{
+			 	$.ajax({
+			        url : 'admin/tiendas/' + id + '/borrar',
+			        type : 'delete',
+			        success:function (data) {
+			        	location.reload();
+			        }
+				});
+			}
+	 	});
 	 });
